refactor: permission

+++ note-service/services/text/business/delete_text.go
@@ -9,8 +9,7 @@ import (
 )
 
 func (biz *business) DeleteText(ctx context.Context, id int) error {
-	// Get media data, without extra infos
-	_, err := biz.textRepo.GetTextById(ctx, id)
+	text, err := biz.textRepo.GetTextById(ctx, id)
 	if err != nil {
 		if err == core.ErrRecordNotFound {
 			return core.ErrNotFound.
@@ -21,6 +20,27 @@ func (biz *business) DeleteText(ctx context.Context, id int) error {
 		return core.ErrInternalServerError.
 			WithError(entity.ErrCannotGetText.Error()).
 			WithDebug(err.Error())
+	}
+
+	requester := core.GetRequester(ctx)
+	uid, _ := core.FromBase58(requester.GetSubject())
+	requesterId := int(uid.GetLocalID())
+
+	note, err := biz.noteRepo.GetNoteById(ctx, int(text.NoteID))
+	if err != nil {
+		if err == core.ErrRecordNotFound {
+			return core.ErrNotFound.
+				WithError(entity.ErrCannotGetText.Error()).
+				WithDebug(err.Error())
+		}
+
+		return core.ErrInternalServerError.
+			WithError(entity.ErrCannotGetText.Error()).
+			WithDebug(err.Error())
+	}
+
+	if requesterId != note.UserId {
+		return core.ErrForbidden.WithError(entity.ErrRequesterIsNotOwner.Error())
 	}
 
 	if err := biz.textRepo.DeleteText(ctx, id); err != nil {
